Feat: 전체 프로젝트 정리 및 Fade-In 추가
