+++ src/app/Dystopia/page.module.css
@@ -18,3 +18,11 @@
   .canvas::-webkit-scrollbar {
     display: none; /* Chrome, Safari, Opera */
   }
+
+  .noSelect {
+    user-select: none;
+    -webkit-user-select: none;
+    -webkit-touch-callout: none;
+    -webkit-tap-highlight-color: transparent;
+    touch-action: manipulation;
+  }
