+++ src/app/Utopia/page.jsx
@@ -1,10 +1,9 @@
 'use client'
 
-import dynamic from 'next/dynamic';
-const Canvas = dynamic(() => import('@react-three/fiber').then((mod) => mod.Canvas), {
-  ssr: false,
-});
 import styles from './page.module.css'
+import { Canvas } from '@react-three/fiber';
+import { useEffect, useRef } from 'react';
+import gsap from 'gsap';
 import Experience from '../../components/Utopia/Experience.jsx';
 import Header from '../Header.jsx';
 //
@@ -12,8 +11,18 @@ import Header from '../Header.jsx';
 //
 export default function Utopia()
 {
-  return (
-    <div style={{ position: 'relative', height: '100vh', width: '100vw' }}>
+const containerRef = useRef();
+
+useEffect(() => {
+  gsap.fromTo(
+    containerRef.current,
+    { opacity: 0 },
+    { opacity: 1, duration: 2, ease: 'power2.inOut' }
+  );
+}, []);
+
+return (
+  <div ref={containerRef} style={{ position: 'relative', height: '100vh', width: '100vw' }}>
       <Header />
       <div className={styles.canvasContainer}>
         <Canvas
