+++ src/components/Utopia/Experience.jsx
@@ -1,36 +1,13 @@
 'use client'
 
-import { useFrame } from '@react-three/fiber'
-import { Text, useGLTF, Stage, Environment, OrbitControls } from '@react-three/drei'
-import { useRef } from 'react'
+import { Text, Stage, Environment, OrbitControls } from '@react-three/drei'
 import { Perf } from 'r3f-perf'
-import * as THREE from 'three'
+import Model from './Model'
 //
 //
 //
 export default function Experience() {
-    const love = useGLTF('./Models/love.glb')
-    const liberation = useGLTF('./Models/liberation.glb')
-    const revolution = useGLTF('./Models/revolution.glb')
-    const resistance = useGLTF('./Models/resistance.glb')
-    const isolation = useGLTF('./Models/isolation.glb')
-    const compliance = useGLTF('./Models/compliance.glb')
-    // console.log(love);
 
-    const loveRef = useRef()
-    const revolutionRef = useRef()
-    const complianceRef = useRef()
-    const resistanceRef = useRef()
-    const isolationRef = useRef()
-    const liberationRef = useRef()
-
-    useFrame((state, delta) => {
-        loveRef.current.rotation.y += delta * 0.2
-        revolutionRef.current.rotation.y += delta * 0.2
-        complianceRef.current.rotation.y += delta * 0.2
-        resistanceRef.current.rotation.y += delta * 0.2
-        liberationRef.current.rotation.y += delta * 0.2
-    })
 
     return (
         <>  
@@ -65,21 +42,8 @@ export default function Experience() {
                 preset="portrait"
                 intensity={ 10 }
             >
-                <primitive ref={loveRef} object={love.scene} position={[20, 12, 0]} scale={5} />
-                <primitive ref={revolutionRef} object={revolution.scene} position={[-18, 12, 0]} scale={3} />
-                <primitive ref={complianceRef} object={compliance.scene} position={[0, 24, 0]} scale={5} />
-                <primitive ref={resistanceRef} object={resistance.scene} position={[0, 0, 0]} scale={5} />
-                <primitive ref={isolationRef} object={isolation.scene} position={[18, 10, 30]} scale={5} />
-                <primitive ref={liberationRef} object={liberation.scene} position={[0, 12, -18]} scale={5} />
-
+                <Model />
             </Stage>
         </>
     )
 }
-
-useGLTF.preload('./Models/love.glb');
-useGLTF.preload('./Models/liberation.glb');
-useGLTF.preload('./Models/revolution.glb');
-useGLTF.preload('./Models/resistance.glb');
-useGLTF.preload('./Models/isolation.glb');
-useGLTF.preload('./Models/compliance.glb');
