+++ src/components/main/shaders/dofPointsMaterial.jsx
@@ -1,6 +1,8 @@
 import * as THREE from 'three'
 import { extend } from '@react-three/fiber'
-
+//
+//
+//
 class DofPointsMaterial extends THREE.ShaderMaterial {
   constructor() {
     super({
