+++ src/components/main/shaders/simulationMaterial.jsx
@@ -1,7 +1,9 @@
 import * as THREE from 'three';
 import { extend } from '@react-three/fiber';
 import { getSphere } from '../../../utils/main/getSphere.jsx';
-
+//
+//
+//
 class SimulationMaterial extends THREE.ShaderMaterial {
   constructor() {
     const positionsTexture = new THREE.DataTexture(
